refactor(DC-704): extract database migrate/seed into DatabaseStartup
